관리자 페이지 관련

+++ Second_Project/src/main/java/org/joonzis/controller/LoginController.java
@@ -685,7 +685,11 @@ public class LoginController {
 	    if (!isValid) {
 	        return ResponseEntity.ok(Map.of("success", false, "message", "인증번호 불일치"));
 	    }
-
+	    
+	    if (passwordEncoder.matches(newPw, user.getUser_pw())) {
+	    	return ResponseEntity.ok(Map.of("success", false, "message", "새 비밀번호가 현재 비밀번호와 동일합니다"));
+		}
+	    System.out.println("현재, 새 비번 다름");
 	    user.setUser_pw(passwordEncoder.encode(newPw));
 	    userservice.updatePassword(user);
 	    userservice.deleteResetToken(token); // 보안 위해 삭제
@@ -838,7 +842,8 @@ public class LoginController {
 	    // 중복 로그인 방지: 이미 로그인 되어있는 세션이 있으면 무효화
         HttpSession existingSession = loginSessions.put(inputId, session);
         if (existingSession != null && !existingSession.isNew()) {
-            return ResponseEntity.status(HttpStatus.CONFLICT).body("이미 로그인된 계정입니다.");
+        	existingSession.invalidate();
+//            return ResponseEntity.status(HttpStatus.CONFLICT).body("이미 로그인된 계정입니다.");
         }
 
 	    session.setAttribute("user_id", inputId);
